requirements.txt
streamlit
polars
Pillow
numpy
colormath
coloraide